用户角色权限模块更新，角色权限模块更新

+++ api/v1/sys_role_api.go
@@ -119,7 +119,7 @@ func (roleApi *RoleApi) UpdateRoleBaseMenu(c *gin.Context) {
 		appRes.Response(http.StatusOK, enum.BIN_JSON_ERROR, nil)
 		return
 	}
-	err = roleService.UpdateRoleMenu(uint(info.RoleId), sysRoleMenuParams.PermissionId)
+	err = roleService.UpdateRoleMenu(info.RoleId, sysRoleMenuParams.PermissionId)
 	if err != nil {
 		return
 	}
